fixup! bench_mutex_pingpong: Add number of clock ticks per cycle

+++ tests/bench_mutex_pingpong/main.c
@@ -82,7 +82,7 @@ int main(void)
     printf("{ \"result\" : %"PRIu32, n);
 #ifdef CLOCK_CORECLOCK
     printf(", \"ticks\" : %"PRIu32,
-           ((TEST_DURATION/US_PER_MS) * (CLOCK_CORECLOCK/KHZ(1)))/n);
+           (uint32_t)((TEST_DURATION/US_PER_MS) * (CLOCK_CORECLOCK/KHZ(1)))/n);
 #endif
     puts(" }");
